migrate to `rug` removed `i8` from `Digit` because `.try_into()` re quieed type annotation cause of things from `rug`

+++ src/algorithm/prime.rs
@@ -174,18 +174,18 @@ pub fn sqrtest(n: isize) -> Result<PrimeStatus, PrimeStatusError> {
 /// ```
 
 pub fn wilson_th(n: isize) -> Result<PrimeStatus, PrimeStatusError> {
-    use num_bigint::BigInt;
+    use rug::Integer;
 
     if n < 2 {
         return Err(PrimeStatusError);
     }
-    fn factorial(n: isize) -> BigInt {
+    fn factorial(n: isize) -> Integer {
         match n {
-            0 | 1 => BigInt::from(1),
-            _ => BigInt::from(n) * factorial(n - 1),
+            0 | 1 => Integer::from(1),
+            _ => Integer::from(n) * factorial(n - 1),
         }
     }
-    if (factorial(n - 1) % BigInt::from(n)) - BigInt::from(n) == BigInt::from(-1) {
+    if (factorial(n - 1) % Integer::from(n)) - Integer::from(n) == Integer::from(-1) {
         Ok(PrimeStatus::Prime)
     } else {
         Ok(PrimeStatus::Composite)
